LPS-179756 Extract CPDefinitionAttachmentsScreenNavigationCategory methods to a CPDefinitionAttachmentsScreenNavigationEntry component and extends CPDefinitionAttachmentsScreenNavigationCategory in it
