+++ modules/apps/commerce/commerce-product-definitions-web/src/main/java/com/liferay/commerce/product/definitions/web/internal/frontend/taglib/servlet/taglib/CPDefinitionAttachmentsScreenNavigationCategory.java
@@ -14,43 +14,14 @@
 
 package com.liferay.commerce.product.definitions.web.internal.frontend.taglib.servlet.taglib;
 
-import com.liferay.commerce.product.configuration.AttachmentsConfiguration;
-import com.liferay.commerce.product.constants.CPActionKeys;
-import com.liferay.commerce.product.constants.CPConstants;
-import com.liferay.commerce.product.ddm.DDMHelper;
-import com.liferay.commerce.product.definitions.web.internal.display.context.CPAttachmentFileEntriesDisplayContext;
-import com.liferay.commerce.product.model.CPDefinition;
-import com.liferay.commerce.product.portlet.action.ActionHelper;
-import com.liferay.commerce.product.service.CPAttachmentFileEntryService;
-import com.liferay.commerce.product.service.CPDefinitionOptionRelService;
 import com.liferay.commerce.product.servlet.taglib.ui.constants.CPDefinitionScreenNavigationConstants;
-import com.liferay.commerce.product.type.CPType;
-import com.liferay.commerce.product.type.CPTypeRegistry;
-import com.liferay.commerce.product.util.CPInstanceHelper;
-import com.liferay.document.library.display.context.DLMimeTypeDisplayContext;
 import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationCategory;
-import com.liferay.frontend.taglib.servlet.taglib.ScreenNavigationEntry;
-import com.liferay.frontend.taglib.servlet.taglib.util.JSPRenderer;
-import com.liferay.item.selector.ItemSelector;
-import com.liferay.portal.configuration.metatype.bnd.util.ConfigurableUtil;
 import com.liferay.portal.kernel.language.Language;
-import com.liferay.portal.kernel.model.User;
-import com.liferay.portal.kernel.security.permission.PermissionChecker;
-import com.liferay.portal.kernel.security.permission.PermissionThreadLocal;
-import com.liferay.portal.kernel.security.permission.resource.PortletResourcePermission;
 import com.liferay.portal.kernel.util.ResourceBundleUtil;
-import com.liferay.portal.kernel.util.WebKeys;
-
-import java.io.IOException;
 
 import java.util.Locale;
-import java.util.Map;
 import java.util.ResourceBundle;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
@@ -58,23 +29,14 @@ import org.osgi.service.component.annotations.Reference;
  * @author Alessio Antonio Rendina
  */
 @Component(
-	configurationPid = "com.liferay.commerce.product.configuration.AttachmentsConfiguration",
-	property = {
-		"screen.navigation.category.order:Integer=70",
-		"screen.navigation.entry.order:Integer=10"
-	},
-	service = {ScreenNavigationCategory.class, ScreenNavigationEntry.class}
+	property = "screen.navigation.category.order:Integer=70",
+	service = ScreenNavigationCategory.class
 )
 public class CPDefinitionAttachmentsScreenNavigationCategory
-	implements ScreenNavigationCategory, ScreenNavigationEntry<CPDefinition> {
+	implements ScreenNavigationCategory {
 
 	@Override
 	public String getCategoryKey() {
-		return CPDefinitionScreenNavigationConstants.CATEGORY_KEY_MEDIA;
-	}
-
-	@Override
-	public String getEntryKey() {
 		return CPDefinitionScreenNavigationConstants.CATEGORY_KEY_MEDIA;
 	}
 
@@ -83,7 +45,7 @@ public class CPDefinitionAttachmentsScreenNavigationCategory
 		ResourceBundle resourceBundle = ResourceBundleUtil.getBundle(
 			"content.Language", locale, getClass());
 
-		return _language.get(
+		return language.get(
 			resourceBundle,
 			CPDefinitionScreenNavigationConstants.CATEGORY_KEY_MEDIA);
 	}
@@ -94,97 +56,7 @@ public class CPDefinitionAttachmentsScreenNavigationCategory
 			SCREEN_NAVIGATION_KEY_CP_DEFINITION_GENERAL;
 	}
 
-	@Override
-	public boolean isVisible(User user, CPDefinition cpDefinition) {
-		if (cpDefinition == null) {
-			return false;
-		}
-
-		PermissionChecker permissionChecker =
-			PermissionThreadLocal.getPermissionChecker();
-
-		CPType cpType = _cpTypeRegistry.getCPType(
-			cpDefinition.getProductTypeName());
-
-		if (_portletResourcePermission.contains(
-				permissionChecker, cpDefinition.getGroupId(),
-				CPActionKeys.VIEW_COMMERCE_PRODUCT_ATTACHMENTS) &&
-			_portletResourcePermission.contains(
-				permissionChecker, cpDefinition.getGroupId(),
-				CPActionKeys.VIEW_COMMERCE_PRODUCT_IMAGES) &&
-			cpType.isMediaEnabled()) {
-
-			return true;
-		}
-
-		return false;
-	}
-
-	@Override
-	public void render(
-			HttpServletRequest httpServletRequest,
-			HttpServletResponse httpServletResponse)
-		throws IOException {
-
-		CPAttachmentFileEntriesDisplayContext
-			cpAttachmentFileEntriesDisplayContext =
-				new CPAttachmentFileEntriesDisplayContext(
-					_actionHelper, _attachmentsConfiguration,
-					_cpAttachmentFileEntryService,
-					_cpDefinitionOptionRelService, _cpInstanceHelper,
-					_ddmHelper, _dlMimeTypeDisplayContext, httpServletRequest,
-					_itemSelector);
-
-		httpServletRequest.setAttribute(
-			WebKeys.PORTLET_DISPLAY_CONTEXT,
-			cpAttachmentFileEntriesDisplayContext);
-
-		_jspRenderer.renderJSP(
-			httpServletRequest, httpServletResponse,
-			"/attachment_file_entries.jsp");
-	}
-
-	@Activate
-	protected void activate(Map<String, Object> properties) {
-		_attachmentsConfiguration = ConfigurableUtil.createConfigurable(
-			AttachmentsConfiguration.class, properties);
-	}
-
 	@Reference
-	private ActionHelper _actionHelper;
-
-	private volatile AttachmentsConfiguration _attachmentsConfiguration;
-
-	@Reference
-	private CPAttachmentFileEntryService _cpAttachmentFileEntryService;
-
-	@Reference
-	private CPDefinitionOptionRelService _cpDefinitionOptionRelService;
-
-	@Reference
-	private CPInstanceHelper _cpInstanceHelper;
-
-	@Reference
-	private CPTypeRegistry _cpTypeRegistry;
-
-	@Reference
-	private DDMHelper _ddmHelper;
-
-	@Reference
-	private DLMimeTypeDisplayContext _dlMimeTypeDisplayContext;
-
-	@Reference
-	private ItemSelector _itemSelector;
-
-	@Reference
-	private JSPRenderer _jspRenderer;
-
-	@Reference
-	private Language _language;
-
-	@Reference(
-		target = "(resource.name=" + CPConstants.RESOURCE_NAME_PRODUCT + ")"
-	)
-	private PortletResourcePermission _portletResourcePermission;
+	protected Language language;
 
 }
